added model loaging and fixed lighting calculations
