+++ README.md
@@ -21,5 +21,6 @@ Actions:
 - Render a 3D model.
 - Load Textures and [sky boxes](https://en.wikipedia.org/wiki/Skybox_(video_games)). 
 - Load shader programs.
+- Load a 3D model.
 - Create the [irradiance map](https://learnopengl.com/PBR/IBL/Diffuse-irradiance) and the [pre-filtered environment map](https://learnopengl.com/PBR/IBL/Specular-IBL) for sky boxes.
 - Render a preview of the position of the lights.
