+++ resources/barrel/Barrel_01.mtl
@@ -1,0 +1,17 @@
+# Blender MTL File: 'Barrel_01.blend'
+# Material Count: 1
+
+newmtl Barrel_01
+Ns 18.000005
+Ka 1.000000 1.000000 1.000000
+Kd 1.000000 1.000000 1.000000
+Ks 0.000000 0.000000 0.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.450000
+d 1.000000
+illum 1
+map_Kd Barrel_01_Textures\\Barrel_01_Radiation\\Barrel_01_Radiation_8-bit_Diffuse.png
+norm Barrel_01_Textures\\Barrel_01_Radiation\\Barrel_01_Radiation_8-bit_Normal.png
+map_Ka Barrel_01_Textures\\Barrel_01_Radiation\\ao.png
+map_Pr Barrel_01_Textures\\Barrel_01_Radiation\\roughness.png
+map_Pm Barrel_01_Textures\\Barrel_01_Radiation\\metallic.png
