+++ resources/helmet/README.md
@@ -1,0 +1,25 @@
+# Battle Damaged Sci-Fi Helmet
+
+## Source
+
+[sketchfab.com/models/b81008d513954189a063ff901f7abfe4](https://sketchfab.com/models/b81008d513954189a063ff901f7abfe4)
+
+[github.com/KhronosGroup/glTF-Sample-Models/tree/master/2.0/DamagedHelmet](https://github.com/KhronosGroup/glTF-Sample-Models/tree/master/2.0/DamagedHelmet)
+
+## Author
+
+[sketchfab.com/theblueturtle_](https://sketchfab.com/theblueturtle_)
+
+## License
+
+Maybe [creativecommons.org/licenses/by-nc/4.0/](https://creativecommons.org/licenses/by-nc/4.0/)
+
+## Preprocessor
+
+[scripts/helmet.py](../../scripts/helmet.py)
+
+## Screenshot
+
+![Screenshot1](screenshot1.png)
+![Screenshot2](screenshot2.png)
+![Screenshot3](screenshot3.png)
