+++ resources/helmet/helmet.mtl
@@ -1,0 +1,17 @@
+# Material Count: 1
+
+newmtl Helmet_01_Mat
+Ns 18.000005
+Ka 1.000000 1.000000 1.000000
+Kd 1.000000 1.000000 1.000000
+Ks 0.000000 0.000000 0.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.450000
+d 1.000000
+illum 1
+map_Kd helmet_basecolor.png
+norm helmet_normal.png
+map_Pm helmet_metalness.png
+map_Ka helmet_occlusion.png
+map_Pr helmet_roughness.png
+map_emissive helmet_emission.tga
