+++ resources/helmet/helmet.scn
@@ -1,0 +1,39 @@
+type: pbrm
+
+lighting:
+    background: 0.15 0.15 0.15
+    environment: workshop
+    skybox: blurred
+    shadow: off
+    ambient: 1.0
+    punctual: 1.0
+
+materials 1:
+    material 0:
+        basecolor_factor: 1 1 1 1
+        metalness_factor: 1
+        roughness_factor: 1
+        basecolor_map: helmet/helmet_basecolor.tga
+        metalness_map: helmet/helmet_metalness.tga
+        roughness_map: helmet/helmet_roughness.tga
+        normal_map: helmet/helmet_normal.tga
+        occlusion_map: helmet/helmet_occlusion.tga
+        emission_map: helmet/helmet_emission.tga
+        double_sided: off
+        enable_blend: off
+        alpha_cutoff: 0
+
+transforms 1:
+    transform 0:
+        1  0  0  0
+        0  1  0  0
+        0  0  1  0
+        0  0  0  1
+
+models 1:
+    model 0:
+        mesh: helmet/helmet.obj
+        skeleton: null
+        attached: -1
+        material: 0
+        transform: 0
